phone field

+++ database/migrations/2020_05_09_023308_create_forms_table.php
@@ -18,7 +18,7 @@ class CreateFormsTable extends Migration
             $table->string('component1')->default('xxx')->nullable();
             $table->string('component2')->default('xxx')->nullable();
             $table->string('component3')->default('xxx')->nullable();
-            $table->integer('component4')->default('055555555')->nullable();
+            $table->bigInteger('component4')->default('055555555')->nullable();
             $table->timestamps();
         });
     }
